// src/models/notification/notification-types.ts
export type AdminNotificationTypes =
	| 'ADMIN_USER_REGISTERED'
	| 'ADMIN_SELLER_APPROVAL'
	| 'ADMIN_NEW_BUYER'
	| 'ADMIN_SINGLE_TURN_IN_REQUEST'
	| 'ADMIN_GROUP_TURN_IN_REQUEST'
	| 'ADMIN_ORDER_DISPATCHED'
	| 'ADMIN_TURN_IN'
	| 'ADMIN_TURN_IN_REJECTED'
	| 'ADMIN_ORDER_CANCELLED'
	| 'ADMIN_ORDER_RETURN_REQUEST'
	| 'ADMIN_ORDER_MATURED'
	| 'ADMIN_PRICE_TABLE_UPDATED'
	| 'ADMIN_PRODUCT_DISABLED'
	| 'ADMIN_PRODUCT_DELETED'
	| 'ADMIN_PRODUCT_ENABLED';

export type SellerNotificationTypes =
	| 'SELLER_ACCOUNT_DISABLED'
	| 'SELLER_ACCOUNT_ENABLED'
	| 'SELLER_ACCOUNT_APPROVED'
	| 'SELLER_PRICE_TABLE_UPDATED'
	| 'SELLER_PRODUCT_DISABLED'
	| 'SELLER_PRODUCT_ENABLED'
	| 'SELLER_GROUP_ORDER_COMPLETED'
	| 'SELLER_GROUP_ORDER_STARTED'
	| 'SELLER_WARNING_ORDER_PLACED'
	| 'SELLER_SINGLE_ORDER_PLACED'
	| 'SELLER_FIRST_ORDER_DELIVERED'
	| 'SELLER_GROUP_ORDER_PLACED'
	| 'SELLER_ORDER_CANCELLED'
	| 'SELLER_RETURN_REQUEST'
	| 'SELLER_RETURN_APPROVED'
	| 'SELLER_RETURN_REJECTED'
	| 'SELLER_RETURN_PICKED_UP'
	| 'SELLER_ORDER_RETURNED'
	| 'SELLER_DELIVERY_NOT_ACCEPTED'
	| 'SELLER_ORDER_DELIVERED'
	| 'SELLER_RIDER_APPROVAL'
	| 'SELLER_REFERRAL_CREATED'
	| 'SELLER_REFERRAL_REWARD'
	| 'SELLER_ORDER_CANCELLED_RIDER'
	| 'SELLER_ORDER_NEW_RIDER_ASSIGN';

export type BuyerNotificationTypes =
	| 'BUYER_ACCOUNT_DISABLED'
	| 'BUYER_ACCOUNT_ENABLED'
	| 'BUYER_GROUP_ORDER_COMPLETE'
	| 'BUYER_ORDER_PLACED'
	| 'BUYER_ORDER_ACCEPTED'
	| 'BUYER_ORDER_REJECTED'
	| 'BUYER_ORDER_CANCELLED'
	| 'BUYER_ORDER_DISPATCHED'
	| 'BUYER_RETURN_REQUEST'
	| 'BUYER_RETURN_REQUEST_APPROVED'
	| 'BUYER_RETURN_REQUEST_REJECTED'
	| 'BUYER_RETURN_COMPLETED'
	| 'BUYER_DELIVERY_NOT_ACCEPTED'
	| 'BUYER_ORDER_DELIVERED'
	| 'BUYER_FIRST_ORDER_DELIVERED'
	| 'BUYER_ORDER_DELIVERED_IN_15_MINUTES'
	| 'BUYER_SIGNUP'
	| 'BUYER_REFERRAL_REWARD'
	| 'BUYER_WALLET_CREATED'
	| 'BUYER_REFUND_INITIATED'
	| 'BUYER_ORDER_ARRIVED'
	| 'BUYER_CASHBACK_RECIEVED';

export type RiderNotificationTypes =
	| 'RIDER_ORDER_RECEIVED'
	| 'RIDER_ORDER_CANCELLED'
	| 'RIDER_ORDER_TRANSFER'
	| 'RIDER_SUBMIT_FLOATING_CASH'
	| 'ON_RIDER_SUBMIT_FLOATING_CASH';

export type NotificationTypes =
	| AdminNotificationTypes
	| SellerNotificationTypes
	| BuyerNotificationTypes
	| RiderNotificationTypes;